# fix: paginate spans by trace start time when rendering span/trace table without filtering