resolves #8420

Adds a specialized pagination optimization for root spans sorted by trace start time, triggered when
- rootSpansOnly: true
- no filter condition
- sort.col: startTime

## Caveats
- Trace-based cursors
    - Cursors track trace positions (rowid + start_time) rather than span positions
- Empty edges with pagination
    - Can return edges: [] while has_next_page: true when traces exist but have no matching root spans
- Per-trace span selection
    - Returns earliest root span per trace when multiple candidates exist

## Tested on PostgreSQL using data with the following stats

- disk: 40.4 GiB
- spans: 18,414,769
- traces: 221,011

## Page Load Time Comparison

### Before
<img width="400" height="343" alt="Screenshot 2025-07-11 at 7 54 11 AM" src="https://github.com/user-attachments/assets/d272b415-d4bf-472a-b35f-3138d3aae2be" />

### After
<img width="400" height="390" alt="Screenshot 2025-07-11 at 7 50 00 AM" src="https://github.com/user-attachments/assets/ff28e9d2-3b50-4f60-aeea-c4d6a35784c5" />

## Sample Queries (PostgreSQL 12.22)

### orphan_span_as_root_span = False
```sql
WITH anon_1 AS (
  SELECT 
    traces.id AS id, 
    traces.start_time AS start_time 
  FROM 
    traces 
  WHERE 
    traces.project_rowid = 1 
    AND traces.start_time >= '2025-06-11 06:00:00+00:00' 
  ORDER BY 
    traces.start_time DESC, 
    traces.id DESC 
  LIMIT 
    31
) 
SELECT 
  DISTINCT ON (anon_1.start_time, anon_1.id) anon_1.id, 
  anon_1.start_time, 
  spans.id AS id_1, 
  spans.start_time AS start_time_1 
FROM 
  anon_1 
  LEFT OUTER JOIN spans ON spans.trace_rowid = anon_1.id 
  AND spans.parent_id IS NULL 
ORDER BY 
  anon_1.start_time DESC, 
  anon_1.id DESC, 
  spans.start_time ASC, 
  spans.id DESC;
```

#### Execution Time: 247.631 ms
```
+-----------------------------------------------------------------------------------------------------------------------------------------------------------+
|QUERY PLAN                                                                                                                                                 |
+-----------------------------------------------------------------------------------------------------------------------------------------------------------+
|Unique  (cost=14287.29..14294.06 rows=31 width=24) (actual time=247.176..247.258 rows=31 loops=1)                                                          |
|  ->  Sort  (cost=14287.29..14289.55 rows=903 width=24) (actual time=247.175..247.247 rows=108 loops=1)                                                    |
|        Sort Key: traces.start_time DESC, traces.id DESC, spans.start_time, spans.id DESC                                                                  |
|        Sort Method: quicksort  Memory: 33kB                                                                                                               |
|        ->  Nested Loop Left Join  (cost=9378.95..14242.96 rows=903 width=24) (actual time=229.193..247.152 rows=108 loops=1)                              |
|              ->  Limit  (cost=9378.51..9382.08 rows=31 width=12) (actual time=228.653..228.739 rows=31 loops=1)                                           |
|                    ->  Gather Merge  (cost=9378.51..24329.20 rows=130006 width=12) (actual time=228.652..228.734 rows=31 loops=1)                         |
|                          Workers Planned: 1                                                                                                               |
|                          Workers Launched: 1                                                                                                              |
|                          ->  Sort  (cost=8378.50..8703.52 rows=130006 width=12) (actual time=221.739..221.742 rows=26 loops=2)                            |
|                                Sort Key: traces.start_time DESC, traces.id DESC                                                                           |
|                                Sort Method: top-N heapsort  Memory: 27kB                                                                                  |
|                                Worker 0:  Sort Method: top-N heapsort  Memory: 27kB                                                                       |
|                                ->  Parallel Seq Scan on traces  (cost=0.00..4508.10 rows=130006 width=12) (actual time=0.058..216.173 rows=110452 loops=2)|
|                                      Filter: ((start_time >= '2025-06-11 06:00:00+00'::timestamp with time zone) AND (project_rowid = 1))                 |
|                                      Rows Removed by Filter: 53                                                                                           |
|              ->  Index Scan using ix_spans_trace_rowid on spans  (cost=0.44..156.50 rows=29 width=16) (actual time=0.265..0.593 rows=3 loops=31)          |
|                    Index Cond: (trace_rowid = traces.id)                                                                                                  |
|                    Filter: (parent_id IS NULL)                                                                                                            |
|                    Rows Removed by Filter: 51                                                                                                             |
|Planning Time: 10.620 ms                                                                                                                                   |
|Execution Time: 247.631 ms                                                                                                                                 |
+-----------------------------------------------------------------------------------------------------------------------------------------------------------+
```

### orphan_span_as_root_span = True
```sql
WITH anon_1 AS (
  SELECT 
    traces.id AS id, 
    traces.start_time AS start_time 
  FROM 
    traces 
  WHERE 
    traces.project_rowid = 1 
    AND traces.start_time >= '2025-06-11 06:00:00+00:00' 
  ORDER BY 
    traces.start_time DESC, 
    traces.id DESC 
  LIMIT 
    31
) 
SELECT 
  DISTINCT ON (anon_1.start_time, anon_1.id) anon_1.id, 
  anon_1.start_time, 
  spans.id AS id_1, 
  spans.start_time AS start_time_1 
FROM 
  anon_1 
  LEFT OUTER JOIN spans ON spans.trace_rowid = anon_1.id 
  AND (
    spans.parent_id IS NULL 
    OR NOT (
      EXISTS (
        SELECT 
          * 
        FROM 
          (
            SELECT 
              spans.span_id AS span_id 
            FROM 
              spans
          ) AS parent_spans 
        WHERE 
          spans.parent_id = parent_spans.span_id
      )
    )
  ) 
ORDER BY 
  anon_1.start_time DESC, 
  anon_1.id DESC, 
  spans.start_time ASC, 
  spans.id DESC;
```

#### Execution Time: 469.562 ms 
```
+----------------------------------------------------------------------------------------------------------------------------------------------------------------+
|QUERY PLAN                                                                                                                                                      |
+----------------------------------------------------------------------------------------------------------------------------------------------------------------+
|Unique  (cost=176436.34..176510.28 rows=31 width=24) (actual time=213.847..214.196 rows=31 loops=1)                                                             |
|  ->  Sort  (cost=176436.34..176460.99 rows=9858 width=24) (actual time=213.846..214.156 rows=213 loops=1)                                                      |
|        Sort Key: traces.start_time DESC, traces.id DESC, spans.start_time, spans.id DESC                                                                       |
|        Sort Method: quicksort  Memory: 41kB                                                                                                                    |
|        ->  Nested Loop Left Join  (cost=9378.95..175782.41 rows=9858 width=24) (actual time=163.855..213.841 rows=213 loops=1)                                 |
|              ->  Limit  (cost=9378.51..9382.08 rows=31 width=12) (actual time=161.458..161.795 rows=31 loops=1)                                                |
|                    ->  Gather Merge  (cost=9378.51..24329.20 rows=130006 width=12) (actual time=104.694..105.029 rows=31 loops=1)                              |
|                          Workers Planned: 1                                                                                                                    |
|                          Workers Launched: 1                                                                                                                   |
|                          ->  Sort  (cost=8378.50..8703.52 rows=130006 width=12) (actual time=29.349..29.361 rows=16 loops=2)                                   |
|                                Sort Key: traces.start_time DESC, traces.id DESC                                                                                |
|                                Sort Method: top-N heapsort  Memory: 27kB                                                                                       |
|                                Worker 0:  Sort Method: quicksort  Memory: 25kB                                                                                 |
|                                ->  Parallel Seq Scan on traces  (cost=0.00..4508.10 rows=130006 width=12) (actual time=0.027..23.912 rows=110452 loops=2)      |
|                                      Filter: ((start_time >= '2025-06-11 06:00:00+00'::timestamp with time zone) AND (project_rowid = 1))                      |
|                                      Rows Removed by Filter: 53                                                                                                |
|              ->  Index Scan using ix_spans_trace_rowid on spans  (cost=0.44..5364.56 rows=318 width=16) (actual time=0.211..1.676 rows=7 loops=31)             |
|                    Index Cond: (trace_rowid = traces.id)                                                                                                       |
|                    Filter: ((parent_id IS NULL) OR (NOT (SubPlan 1)))                                                                                          |
|                    Rows Removed by Filter: 48                                                                                                                  |
|                    SubPlan 1                                                                                                                                   |
|                      ->  Index Only Scan using uq_spans_span_id on spans spans_1  (cost=0.56..8.58 rows=1 width=0) (actual time=0.032..0.032 rows=1 loops=1580)|
|                            Index Cond: (span_id = (spans.parent_id)::text)                                                                                     |
|                            Heap Fetches: 1475                                                                                                                  |
|Planning Time: 2.124 ms                                                                                                                                         |
|JIT:                                                                                                                                                            |
|  Functions: 22                                                                                                                                                 |
|  Options: Inlining false, Optimization false, Expressions true, Deforming true                                                                                 |
|  Timing: Generation 15.075 ms, Inlining 0.000 ms, Optimization 3.868 ms, Emission 52.969 ms, Total 71.912 ms                                                   |
|Execution Time: 469.562 ms                                                                                                                                      |
+----------------------------------------------------------------------------------------------------------------------------------------------------------------+
```